scrapy/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
Automat==20.2.0
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
cssselect==1.1.0
distlib==0.3.1
Django==3.1.1
django-crispy-forms==1.9.2
django-rest-swagger==2.2.0
django-truncate==0.1
djangorestframework==3.11.1
docutils==0.16
drf-yasg==1.17.1
filelock==3.0.12
httpie==2.2.0
hyperlink==20.0.1
idna==2.10
imagesize==1.2.0
incremental==17.5.0
inflection==0.5.1
isort==5.5.1
itemadapter==0.1.0
itemloaders==1.0.2
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
npm==0.1.1
ObjDict==0.4.4
openapi-codec==1.3.2
optional-django==0.1.0
packaging==20.4
parsel==1.6.0
pdoc3==0.9.1
pluggy==0.13.1
Protego==0.1.16
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
rope==0.17.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Scrapy==2.3.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
swagger-ui-py==0.3.0
toml==0.10.1
tox==3.20.0
Twisted @ https://github.com/twisted/twisted/archive/twisted-18.7.0.zip
uritemplate==3.0.1
urllib3==1.25.10
users==1.0.dev0
virtualenv==20.0.31
w3lib==1.22.0
wrapt==1.12.1
zope.interface==5.1.0
